Refactor code, reformat code, remove comment
