--- src/Configuration/DI/index.ts
@@ -1,1 +1,0 @@
-export * from './DI';
