+++ src/Modules/Station/Components/PlaylistTabs/History/History.tsx
@@ -1,5 +1,5 @@
 import { BaseComponent } from 'BaseComponent';
-import { Inject } from 'Configuration/DI';
+import { Inject } from 'Configuration/DependencyInjector';
 import { Song } from 'Models';
 import * as React from 'react';
 import { SongServices } from 'Services/Http';
