+++ src/Modules/User/Components/Login/ButtonGoogleLogin/ButtonGoogleLogin.tsx
@@ -1,4 +1,4 @@
-import { Inject } from 'Configuration/DI';
+import { Inject } from 'Configuration/DependencyInjector';
 import * as React from 'react';
 import { Component } from 'react';
 import { GoogleLogin } from 'react-google-login';
