+++ src/Services/Http/StationServices/SongServices.ts
@@ -1,4 +1,4 @@
-import { Service } from 'Configuration/DI';
+import { Service } from 'Configuration/DependencyInjector';
 import { Song } from 'Models';
 import { Observable } from 'rxjs/Observable';
 import { HttpServices } from '../HttpServices';
